requirements.txt
google-genai>=0.3.0
pandas>=2.0.0
fhir.resources>=8.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pathlib>=1.0.1